+++ CHANGES.md
@@ -43,6 +43,7 @@ Release Notes.
 * Add `Message Queue Consuming Count` metric for MQ consuming service and endpoint.
 * Add `Message Queue Avg Consuming Latency` metric for MQ consuming service and endpoint.
 * Support `-Inf` as bucket in the meter system.
+* Support search browser service.
 
 #### UI